fix(dashboard): delete territory feature, remove unused
