+++ api/src/pdc/services/dashboard/DashboardServiceProvider.ts
@@ -15,9 +15,6 @@ import { DeleteOperatorAction } from "@/pdc/services/dashboard/actions/operators
 import { OperatorAction } from "@/pdc/services/dashboard/actions/operators/OperatorAction.ts";
 import { OperatorsAction } from "@/pdc/services/dashboard/actions/operators/OperatorsAction.ts";
 import { UpdateOperatorAction } from "@/pdc/services/dashboard/actions/operators/UpdateOperatorAction.ts";
-import { DeleteTerritoryAction } from "@/pdc/services/dashboard/actions/territories/DeleteTerritoryAction.ts";
-import { TerritoriesAction } from "@/pdc/services/dashboard/actions/territories/TerritoriesAction.ts";
-import { TerritoryAction } from "@/pdc/services/dashboard/actions/territories/TerritoryAction.ts";
 import { CreateUserAction } from "@/pdc/services/dashboard/actions/users/CreateUserAction.ts";
 import { DeleteUserAction } from "@/pdc/services/dashboard/actions/users/DeleteUserAction.ts";
 import { UpdateUserAction } from "@/pdc/services/dashboard/actions/users/UpdateUserAction.ts";
@@ -26,7 +23,6 @@ import { UsersAction } from "@/pdc/services/dashboard/actions/users/UsersAction.
 import { CampaignsRepository } from "@/pdc/services/dashboard/providers/CampaignsRepository.ts";
 import { JourneysRepository } from "@/pdc/services/dashboard/providers/JourneysRepository.ts";
 import { OperatorsRepository } from "@/pdc/services/dashboard/providers/OperatorsRepository.ts";
-import { TerritoriesRepository } from "@/pdc/services/dashboard/providers/TerritoriesRepository.ts";
 import { UsersRepository } from "@/pdc/services/dashboard/providers/UsersRepository.ts";
 
 /* eslint-enable */
@@ -38,7 +34,6 @@ import { UsersRepository } from "@/pdc/services/dashboard/providers/UsersReposit
     JourneysRepository,
     CampaignsRepository,
     UsersRepository,
-    TerritoriesRepository,
   ],
   handlers: [
     JourneysOperatorsByMonthAction,
@@ -47,9 +42,6 @@ import { UsersRepository } from "@/pdc/services/dashboard/providers/UsersReposit
     JourneysIncentiveByDayAction,
     CampaignsAction,
     CampaignApdfAction,
-    TerritoryAction,
-    TerritoriesAction,
-    DeleteTerritoryAction,
     OperatorAction,
     OperatorsAction,
     CreateOperatorAction,
