--- api/src/pdc/services/dashboard/dto/Territories.ts
@@ -1,20 +1,0 @@
-import { Infer, object, optional } from "@/lib/superstruct/index.ts";
-import { Id, Siret, TerritoryType, Varchar } from "@/pdc/providers/superstruct/shared/index.ts";
-
-export const Territories = object({
-  id: optional(Id),
-  page: optional(Id),
-  limit: optional(Id),
-});
-export const CreateTerritory = object({
-  name: Varchar,
-  type: TerritoryType,
-  siret: Siret,
-});
-export const DeleteTerritory = object({
-  id: Id,
-});
-
-export type Territories = Infer<typeof Territories>;
-export type DeleteTerritory = Infer<typeof DeleteTerritory>;
-export type CreateTerritory = Infer<typeof CreateTerritory>;
