+++ api/src/pdc/services/territory/TerritoryServiceProvider.ts
@@ -3,6 +3,8 @@ import { ServiceProvider as AbstractServiceProvider } from "@/ilos/core/index.ts
 import { defaultMiddlewareBindings } from "@/pdc/providers/middleware/index.ts";
 import { ValidatorMiddleware as SuperStructValidator } from "@/pdc/providers/superstruct/ValidatorMiddleware.ts";
 import { ValidatorExtension, ValidatorMiddleware } from "@/pdc/providers/validator/index.ts";
+import { DeleteTerritoryAction } from '@/pdc/services/territory/actions/group/DeleteTerritoryAction.ts';
+import { DeleteTerritoryActionV2 } from "@/pdc/services/territory/actions/group/DeleteTerritoryActionV2.ts";
 import { ListTerritoryActionV2 } from "@/pdc/services/territory/actions/group/ListTerritoryActionV2.ts";
 import { create } from "@/pdc/services/territory/contracts/create.schema.ts";
 import { deleteTerritory } from "@/pdc/services/territory/contracts/delete.schema.ts";
@@ -49,12 +51,13 @@ import { TerritoryRepositoryProvider } from "./providers/TerritoryRepositoryProv
     FindTerritoryAction,
     ListTerritoryAction,
     ListGeoAction,
+    DeleteTerritoryAction,
+    DeleteTerritoryActionV2,
     UpdateTerritoryAction,
     PatchContactsTerritoryAction,
     CreateTerritoryAction,
     FindGeoBySirenAction,
     ListTerritoryActionV2,
-    //CreateTerritoryActionV2,
     GetAuthorizedCodesAction,
     IndexAllGeoAction,
   ],
