+++ api/src/pdc/services/territory/actions/group/DeleteTerritoryAction.ts
@@ -1,7 +1,7 @@
 import { handler, KernelInterfaceResolver } from "@/ilos/common/index.ts";
 import { Action as AbstractAction } from "@/ilos/core/index.ts";
-import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
+import { hasPermissionMiddleware } from "@/pdc/providers/middleware/middlewares.ts";
 import { handlerConfig, ParamsInterface, ResultInterface } from "@/pdc/services/territory/contracts/delete.contract.ts";
 import { alias } from "@/pdc/services/territory/contracts/delete.schema.ts";
 import { TerritoryRepositoryProviderInterfaceResolver } from "../../interfaces/TerritoryRepositoryProviderInterface.ts";
@@ -21,7 +21,7 @@ export class DeleteTerritoryAction extends AbstractAction {
     super();
   }
 
-  public async handle(params: ParamsInterface): Promise<ResultInterface> {
+  public override async handle(params: ParamsInterface): Promise<ResultInterface> {
     await this.territoryRepository.delete(params._id);
     await this.kernel.call(
       "user:deleteAssociated",
